+++ core/baykit/bayserver/bcf/bcf_key_val.py
@@ -1,9 +1,0 @@
-from baykit.bayserver.bcf.bcf_object import BcfObject
-
-class BcfKeyVal(BcfObject):
-
-    def __init__(self, key, val, file_name, line_no):
-        super().__init__(file_name, line_no)
-        self.key = key
-        self.value = val
-