+++ core/baykit/bayserver/docker/permission.py
@@ -1,4 +1,0 @@
-from baykit.bayserver.docker.docker import Docker
-
-class Permission(Docker):
-    pass