+++ core/baykit/bayserver/protocol/command_unpacker.py
@@ -1,7 +1,0 @@
-from abc import ABCMeta, abstractmethod
-
-from baykit.bayserver.util.reusable import Reusable
-
-class CommandUnPacker(Reusable, metaclass=ABCMeta):
-    def packet_received(self, pkt):
-        pass