+++ core/baykit/bayserver/protocol/packet_unpacker.py
@@ -1,9 +1,0 @@
-from abc import ABCMeta, abstractmethod
-
-from baykit.bayserver.util.reusable import Reusable
-
-class PacketUnPacker(Reusable, metaclass=ABCMeta):
-
-    @abstractmethod
-    def bytes_received(self, bytes):
-        pass