+++ docker/ajp/baykit/bayserver/docker/ajp/ajp_docker.py
@@ -1,4 +1,0 @@
-from baykit.bayserver.docker.docker import Docker
-
-class AjpDocker(Docker):
-    PROTO_NAME = "ajp"