+++ docker/http/baykit/bayserver/docker/http/h1/h1_command.py
@@ -1,4 +1,0 @@
-from baykit.bayserver.protocol.command import Command
-
-class H1Command(Command):
-    pass