+++ stage/bin/bayserver.sh
@@ -9,14 +9,10 @@ for arg in $args; do
   fi
 done
 
-pycmd=python
-pyver=`python --version 2>&1 | sed 's/Python //' | sed 's/\..*//'`
-if [ "$pyver" != "3" ]; then
-   pycmd=python3
-fi
-
+site=${base}/../site-packages
+export PYTHONPATH=${site}
 if [ "$daemon" = 1 ]; then
-   $pycmd $base/bootstrap.py $* < /dev/null  > /dev/null 2>&1 &
+   ${site}/bin/bayserver_py $* < /dev/null  > /dev/null 2>&1 &
 else
-   $pycmd $base/bootstrap.py $* 
+   ${site}/bin/bayserver_py $* 
 fi