+++ NEWS.md
@@ -1,6 +1,16 @@
 # BayServer for Python
 
 
+# 2.2.0
+
+- Supports pip install
+
+# 2.1.0
+
+- Supports multi core mode for Windows
+- Change name of banjo docker to "maccaferri" docker
+- Fixes some bugs
+
 # 2.0.3
 
 - Fixes Problem of POST request which causes 404